Restructured Docker files.
Docker relies on vanilla linux images now (clusterbase for Harvard,
ubuntu for everything else).
Added a development image.
Removed the necessity to build ubench SWIG interfaces in order to run
tools. This was causing problems because the TF-0.12 pip package does
not come with Eigen sources, so we can't build the microbenchmarks.

We'll need to update TF, but this provides a more sane path to do so.