Major change to Model interface.
The old model interface was overloaded and oversimplistic.
I've scrapped the old Model and added a set of four new classes:
one abstract class:
BaseModel: now is the root of all Models; it is just used for subclass checks
and three concrete classes, which implement increasing levels of functionality:
ImmutableModel: a model that can't be run or changed, but can be analyzed
StaticModel: a model that can't be run, but can be analyzed and updated
DynamicModel: a model that can be analyzed, updated, and evaluated

NOTE: the changes to the inheritence structure surrounding BaseModel breaks a
good chunk of the TFFramework's loading methods. This is okay, as those methods
needed to get torn out, segregated into a separate set of classes, and rewritten
anyways.
HOWEVER, it also means that most of the tests no longer pass in this commit.
I believe all of the failures halt on TFTramework's multiple models assertion.
While there are almost certainly other bugs lurking, I think I've cleaned up
everything that doesn't depend on actually being able to poke at a model.