Added Loaders.
Loaders are flexible methods for ingesting model implementations.
Instead of having one all-encompassing mechanism, loaders provide a way of
managing a variety of (perhaps mutually-incompatible) methods for doing so.
In addition, they can be provided for only certain frameworks.

Tests are not all updated, but test_tf and (the new) test_tf_loaders both pass.