Internalized all Dnnamo tools.
The old tools were a bit awkward to work with, since they were separate and,
honestly, worse than the rest of Dnnamo.
I've moved them inside the library and replaced the run.py file with a __main__
file. So instead of running:
  ./tools/run.py ...
You now run:
  python -m dnnamo ...
It works just like before, except it's way cleaner from my side.
Also, I removed the crufty everything-is-a-Tool-class and the use of imp to
load them all. That was a bad idea before, and it's a bad idea now.