Expanded datatags and framework interface.
Originally, there were just a few framework accessors.
Then we added datatags for caching and added a couple accessors.
Then we added modes for training and inference sides of a model.
Then we added static and dynamic graphs because of framework JIT optimization.
The total number of things and cases started to get unwieldy.

So I've reimplemented the interface a bit, such that there are four categories
of accessors:
  graph
  weights
  timing
  ivalues
And each of these now take a variety of flags:
  mode: training, inference
  scope: static, dynamic
  ops: native, primitive
The Datatag class had to be completely redone (otherwise I was looking at
  manually tracking at least 16 individual and inconsistent values).
I've changed datatags into augmented tuples, and these are used to drive the
  underlying data manager's cache. They're also used for invalidation, and
  as such, they support masking.

The framework interface got a bit simpler on the accessor side, and the data
  collection side got unified into a single "collect" method which takes a
  datatag as an argument. This is a much harder function to write, and
  overriding it in frameworks will be harder, but it was going to happen
  anyways, due to the increased number of pieces of data we're collecting.

This breaks a bunch of stuff, but I plan on breaking a lot more with the
  generic graph and generic op changes.