Added generic op, tensor, and graph classes.
Using the native forms of graphs was problematic for Frameworks.
In moving over to a Dnnamo-specific generic representation for graphs,
we also needed generic forms of operations. This wasn't so bad, since we
basically abstracted the primop and absgraph classes into something more
flexible. While doing so, I also added tensors as a first-class object.
This added a bit of work, but it simplified the interface for getting at
tensors. This makes certain types of analysis dramatically easier, as well
as giving us a straightforward solution for what to return from ivalues().

Tests are included for many parts of this.