Fixed datatag collector inheritance problems.
It turns out that the inheritance scheme I concocted was only working
one direction. I.E., the subclass (TFFramework) was getting all of the
default collector methods in its parent (Framework), but the parent was
also getting all of the specialized methods in its child. This meant
that all Framework subclasses were getting their siblings' methods, which
of course, started referencing functions that didn't exist. (For
instance, the fused rungraph/timing collector referenced a private
runmetadata-parsing method.)

This has been fixed with a set of external decorator classes. It's not
quite as elegant as I'd like, but it works correctly.

Two warnings:
1. If you subclass Framework, you *MUST* also create a subclass of the
  _collector decorator class, and it *MUST* copy the registry class
  attribute. The copy provides unique storage for each collector,
  enabling inheritance.
2. Framework subclasses must also have a _collector_registry class
  attribute which points to the decorator subclass created in step 1.
  This lets the subclass find the appropriate set of collector methods.

These aren't hard, and I feel like the trade-off in ease of use for
actually declaring and managing collector functions is worth it, but
it's something to be aware of, especially for test cases that create
temporary subclasses of Framework---these need to abide by the above
rules if they want to collect data correctly.