Added a generic Registry class for Dnnamo.
There's been a few of these cropping up, and I thought it was high time to have
  a single, shared implementation for it.
NOTE: I've made it center around a bijective map. So 1-1 registries only.
  (I think this is the behavior of all intended instances of it)
I've added test cases for the Registry class. They should be expanded.
I've added __delitem__ support for Bimaps. This deletes an entry from both the
  forward and backward maps when one is deleted (e.g., 'del bimap.l')

PrimopTypes has been moved to a Regstiry subclass.
ToolRegistry has NOT yet been moved, but it should be.