Updated Estimators to be single-op instead of collections.
I realized that it was impossible to have several different regression tactics
depending on the primop. (So you couldn't, for instance, use an OLS regressor
and an SVM in the same model.) I'll add an EstimatorGroup to collect different
Estimators into a single object later.

Also added I/O to Estimator objects.